refactor: align internal precision to exchange native (SZSE px1e4/qty1e2, SSE px1e3/qty1e3); drop __int128 for int64
- Internal price/qty precision now equals the exchange native precision (SZSE px x1e4, qty x1e2; SSE px x1e3, qty x1e3), so raw->internal is mul=1 (identity, zero loss).
- Amount = price x qty: product precision is x1e6 (both exchanges); divide by per-exchange factor (SZSE /100, SSE /10) to land on exchange amount precision. Product no longer overflows int64, so all __int128 removed (BidWeightValue/AskWeightValue/TotalValueTrade now int64). This also lets the engine compile on MSVC (which lacks __int128).
- tick1Cent() returns 100 (SZSE) / 10 (SSE) in native price precision; CYB cage keeps TICK_1CENT=100 (cage is SZSE-only).
- Build validated with MSVC + GoogleTest (33/33 tests pass); MSVC can now compile all engine targets.

+++ cpp v2/behave/axob.h
@@ -105,15 +105,16 @@ public:
 #endif
 
     // ==================== 加权统计 ====================
-    // *Value 为全簿 Σ(price × qty) 累计和: 统一定点 ×10^5 下单档乘积即可达 1e21,
-    // 全簿求和更大, 必须 __int128 (实测 000001 买侧 1e8 股 × 10.54 元 = 1.05e19 已越 int64)。
+    // *Value 为全簿 Σ(price × qty) 累计和 (产品精度 ×10^6, 即 价格×数量乘积精度)。
+    // 内部精度对齐交易所原生 (深: 价×10^4/量×10^2, 沪: 价×10^3/量×10^3), 乘积精度 ×10^6,
+    // 对境内任何真实订单簿均远低于 int64 上限, 故用 int64, 无需 __int128。
     // *Size 为纯数量累计, int64 足够。
     int64_t  BidWeightSize    = 0;
-    __int128 BidWeightValue   = 0;
+    int64_t  BidWeightValue   = 0;
     int64_t  AskWeightSize    = 0;
-    __int128 AskWeightValue   = 0;
+    int64_t  AskWeightValue   = 0;
     int64_t  AskWeightSizeEx  = 0;
-    __int128 AskWeightValueEx = 0;
+    int64_t  AskWeightValueEx = 0;
 
     // ==================== 缓存单 ====================
     // [v2优化] 从 unique_ptr 改为栈上对象，省掉堆分配（ObOrder 仅 32B）
